+++ scripts/generate-deployment_yaml.py
@@ -1,4 +1,5 @@
 import logging
+import sqlalchemy
 
 import esdglider.config as config
 
@@ -25,9 +26,10 @@ if __name__ == "__main__":
 
     with open("db/glider-db-prod.txt", "r") as f:
         conn_string = f.read()
+    engine = sqlalchemy.create_engine(conn_string)
 
     config.make_deployment_yaml(
+        engine=engine, 
         deployment_name="amlr04-20231128",
         out_path=path_config,
-        db_url=conn_string,
     )