+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Changed package config fully to pyproject.toml
 - Added a new helper function to handle the generation of paths strictly within the glider deployment folder out into its own function: `glider.get_path_deployment_glider`. This function is called by `glider.get_path_deployment`
+- Changed all config module functions to accept an engine argument, rather than a db_url
 
 
 ## [0.2.0] - 2025-07-22